refactor: initial implementation of separeted lib and bin
