+++ Cargo.toml
@@ -1,88 +1,7 @@
-[package]
-name = "rpxy"
-version = "0.3.0"
-authors = ["Jun Kurihara"]
-homepage = "https://github.com/junkurihara/rust-rpxy"
-repository = "https://github.com/junkurihara/rust-rpxy"
-license = "MIT"
-readme = "README.md"
-edition = "2021"
-publish = false
+[workspace]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[features]
-default = ["http3", "sticky-cookie"]
-http3 = ["quinn", "h3", "h3-quinn"]
-sticky-cookie = ["base64", "sha2", "chrono"]
-
-[dependencies]
-anyhow = "1.0.72"
-clap = { version = "4.3.17", features = ["std", "cargo", "wrap_help"] }
-rand = "0.8.5"
-toml = { version = "0.7.6", default-features = false, features = ["parse"] }
-rustc-hash = "1.1.0"
-serde = { version = "1.0.174", default-features = false, features = ["derive"] }
-bytes = "1.4.0"
-thiserror = "1.0.44"
-x509-parser = "0.15.0"
-derive_builder = "0.12.0"
-futures = { version = "0.3.28", features = ["alloc", "async-await"] }
-tokio = { version = "1.29.1", default-features = false, features = [
-  "net",
-  "rt-multi-thread",
-  "time",
-  "sync",
-  "macros",
-] }
-async-trait = "0.1.72"
-hot_reload = "0.1.2" # reloading certs
-
-# http and tls
-hyper = { version = "0.14.27", default-features = false, features = [
-  "server",
-  "http1",
-  "http2",
-  "stream",
-] }
-hyper-rustls = { version = "0.24.1", default-features = false, features = [
-  "tokio-runtime",
-  "webpki-tokio",
-  "http1",
-  "http2",
-] }
-tokio-rustls = { version = "0.24.1", features = ["early-data"] }
-rustls-pemfile = "1.0.3"
-rustls = { version = "0.21.5", default-features = false }
-webpki = "0.22.0"
-
-# logging
-tracing = { version = "0.1.37" }
-tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
-
-# http/3
-# quinn = { version = "0.9.3", optional = true }
-quinn = { path = "./quinn/quinn", optional = true } # Tentative to support rustls-0.21
-h3 = { path = "./h3/h3/", optional = true }
-# h3-quinn = { path = "./h3/h3-quinn/", optional = true }
-h3-quinn = { path = "./h3-quinn/", optional = true } # Tentative to support rustls-0.21
-
-# cookie handling for sticky cookie
-chrono = { version = "0.4.26", default-features = false, features = [
-  "unstable-locales",
-  "alloc",
-  "clock",
-], optional = true }
-base64 = { version = "0.21.2", optional = true }
-sha2 = { version = "0.10.7", default-features = false, optional = true }
-
-
-[target.'cfg(not(target_env = "msvc"))'.dependencies]
-tikv-jemallocator = "0.5.0"
-
-
-[dev-dependencies]
-
+members = ["rpxy-bin", "rpxy-lib"]
+exclude = ["quinn", "h3-quinn", "h3"]
 
 [profile.release]
 codegen-units = 1
