+++ rpxy-bin/src/constants.rs
@@ -1,0 +1,2 @@
+pub const LISTEN_ADDRESSES_V4: &[&str] = &["0.0.0.0"];
+pub const LISTEN_ADDRESSES_V6: &[&str] = &["[::]"];
