+++ rpxy-bin/src/error.rs
@@ -1,0 +1,1 @@
+pub use anyhow::{anyhow, bail, ensure, Context};
