+++ rpxy-lib/Cargo.toml
@@ -1,0 +1,77 @@
+[package]
+name = "rpxy-lib"
+version = "0.4.0"
+authors = ["Jun Kurihara"]
+homepage = "https://github.com/junkurihara/rust-rpxy"
+repository = "https://github.com/junkurihara/rust-rpxy"
+license = "MIT"
+readme = "README.md"
+edition = "2021"
+publish = false
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[features]
+default = ["http3", "sticky-cookie"]
+http3 = ["quinn", "h3", "h3-quinn"]
+sticky-cookie = ["base64", "sha2", "chrono"]
+
+[dependencies]
+rand = "0.8.5"
+rustc-hash = "1.1.0"
+bytes = "1.4.0"
+derive_builder = "0.12.0"
+futures = { version = "0.3.28", features = ["alloc", "async-await"] }
+tokio = { version = "1.29.1", default-features = false, features = [
+  "net",
+  "rt-multi-thread",
+  "time",
+  "sync",
+  "macros",
+] }
+async-trait = "0.1.72"
+hot_reload = "0.1.2" # reloading certs
+
+# Error handling
+anyhow = "1.0.72"
+thiserror = "1.0.44"
+
+# http and tls
+hyper = { version = "0.14.27", default-features = false, features = [
+  "server",
+  "http1",
+  "http2",
+  "stream",
+] }
+hyper-rustls = { version = "0.24.1", default-features = false, features = [
+  "tokio-runtime",
+  "webpki-tokio",
+  "http1",
+  "http2",
+] }
+tokio-rustls = { version = "0.24.1", features = ["early-data"] }
+rustls = { version = "0.21.5", default-features = false }
+webpki = "0.22.0"
+x509-parser = "0.15.0"
+
+# logging
+tracing = { version = "0.1.37" }
+
+# http/3
+# quinn = { version = "0.9.3", optional = true }
+quinn = { path = "../quinn/quinn", optional = true } # Tentative to support rustls-0.21
+h3 = { path = "../h3/h3/", optional = true }
+# h3-quinn = { path = "./h3/h3-quinn/", optional = true }
+h3-quinn = { path = "../h3-quinn/", optional = true } # Tentative to support rustls-0.21
+
+# cookie handling for sticky cookie
+chrono = { version = "0.4.26", default-features = false, features = [
+  "unstable-locales",
+  "alloc",
+  "clock",
+], optional = true }
+base64 = { version = "0.21.2", optional = true }
+sha2 = { version = "0.10.7", default-features = false, optional = true }
+
+
+[dev-dependencies]
