+++ rpxy-lib/src/utils/bytes_name.rs
@@ -23,6 +23,9 @@ impl PathNameBytesExp {
   pub fn len(&self) -> usize {
     self.0.len()
   }
+  pub fn is_empty(&self) -> bool {
+    self.0.len() == 0
+  }
   pub fn get<I>(&self, index: I) -> Option<&I::Output>
   where
     I: std::slice::SliceIndex<[u8]>,
